# PARQUET-625: Improve RLE read performance