Also fixes the RLE decoding benchmark which did not pass the size to the
buffer.