@xhochy 2016-05-29
## Level-Benchmark

TL;DR: In the best case 80x faster, in a not so optimal case still >2x

### Before

```
Benchmark                Time(ns)    CPU(ns) Iterations
-------------------------------------------------------
BM_RleEncoding/1024/1        1737       1716     408325  1.11169GB/s   569.187M items/s
BM_RleEncoding/1024/8        4299       4260     164464  458.532MB/s   229.266M items/s
BM_RleEncoding/1024/16       2594       2590     270794  754.116MB/s   377.058M items/s
BM_RleEncoding/4k/1          6737       6662     105396  1.14521GB/s   586.346M items/s
BM_RleEncoding/4k/8         16864      16853      41642  463.558MB/s   231.779M items/s
BM_RleEncoding/4k/16        10246      10122      65130  771.811MB/s   385.906M items/s
BM_RleEncoding/32k/1        53452      53316      13157  1.14478GB/s   586.126M items/s
BM_RleEncoding/32k/8       136162     134269       5273  465.485MB/s   232.743M items/s
BM_RleEncoding/32k/16       80708      80663       8709  774.824MB/s   387.412M items/s
BM_RleEncoding/64k/1       108464     106531       6318  1.14586GB/s   586.681M items/s
BM_RleEncoding/64k/8       267363     266153       2587  469.654MB/s   234.827M items/s
BM_RleEncoding/64k/16      162001     159799       4367  782.234MB/s   391.117M items/s
BM_RleDecoding/1024/1        5548       5483     127444  356.197MB/s   178.099M items/s
BM_RleDecoding/1024/8        8006       7986      87942  244.579MB/s   122.289M items/s
BM_RleDecoding/1024/16       4082       4032     172331  484.397MB/s   242.198M items/s
BM_RleDecoding/4k/1         21905      21898      32156  356.764MB/s   178.382M items/s
BM_RleDecoding/4k/8         32025      31899      21903  244.917MB/s   122.459M items/s
BM_RleDecoding/4k/16        16089      15898      44317  491.407MB/s   245.703M items/s
BM_RleDecoding/32k/1       181379     179726       3937  347.751MB/s   173.875M items/s
BM_RleDecoding/32k/8       259814     259150       2710  241.173MB/s   120.587M items/s
BM_RleDecoding/32k/16      129595     128061       5479  488.049MB/s   244.024M items/s
BM_RleDecoding/64k/1       359974     356746       1955  350.389MB/s   175.195M items/s
BM_RleDecoding/64k/8       523461     517575       1350  241.511MB/s   120.755M items/s
BM_RleDecoding/64k/16      255580     255392       2742  489.444MB/s   244.722M items/s
```

### After

``` 
Benchmark                Time(ns)    CPU(ns) Iterations
-------------------------------------------------------
BM_RleEncoding/1024/1        1776       1751     397093  1115.15MB/s   557.577M items/s
BM_RleEncoding/1024/8        4388       4355     161293  448.483MB/s   224.242M items/s
BM_RleEncoding/1024/16       2651       2614     270514  747.249MB/s   373.625M items/s
BM_RleEncoding/4k/1          7063       6967      97718  1121.36MB/s   560.679M items/s
BM_RleEncoding/4k/8         17405      17202      40542  454.154MB/s   227.077M items/s
BM_RleEncoding/4k/16         9856       9854      70435  792.829MB/s   396.415M items/s
BM_RleEncoding/32k/1        54432      53655      13081  1.13754GB/s   582.421M items/s
BM_RleEncoding/32k/8       133123     131898       5328  473.852MB/s   236.926M items/s
BM_RleEncoding/32k/16       78815      77458       9045  806.893MB/s   403.447M items/s
BM_RleEncoding/64k/1       112882     112799       6506  1108.16MB/s   554.082M items/s
BM_RleEncoding/64k/8       264649     264508       2649  472.576MB/s   236.288M items/s
BM_RleEncoding/64k/16      159566     157676       4509  792.765MB/s   396.383M items/s
BM_RleDecoding/1024/1         123        122    5678730  15.5707GB/s   7.78537G items/s
BM_RleDecoding/1024/8        3431       3356     208913  582.013MB/s   291.007M items/s
BM_RleDecoding/1024/16       1832       1811     386572  1078.32MB/s   539.158M items/s
BM_RleDecoding/4k/1           241        241    2962875  31.6466GB/s   15.8233G items/s
BM_RleDecoding/4k/8         13812      13480      50581  579.573MB/s   289.787M items/s
BM_RleDecoding/4k/16         6948       6945     100030  1124.86MB/s   562.429M items/s
BM_RleDecoding/32k/1         1948       1926     361703  31.6953GB/s   15.8476G items/s
BM_RleDecoding/32k/8       106114     105820       6649  590.625MB/s   295.312M items/s
BM_RleDecoding/32k/16       55898      55121      12749   1.1073GB/s   566.935M items/s
BM_RleDecoding/64k/1         4011       3842     182498   31.773GB/s   15.8865G items/s
BM_RleDecoding/64k/8       211715     211433       3300  591.203MB/s   295.601M items/s
BM_RleDecoding/64k/16      113640     110968       6368  1.10005GB/s   563.225M items/s
```