@wesm 2016-06-03
There's a compiler warning failing the build, also may need rebase