@wesm 2016-06-03
+1, thank you for this perf optimization!