@wesm 2016-06-03
There seems to be some sluggishness in ASF infra; this is merged to ASF git but the commit hasn't been mirrored yet to github